+++ interfaces/portal/src/app/domains/fsp-account-management/intersolve-visa.helper.ts
@@ -1,0 +1,12 @@
+import { VisaCard121Status } from '@121-service/src/fsp-integrations/integrations/intersolve-visa/enums/wallet-status-121.enum';
+
+export const VISA_CARD_STATUS_LABELS: Record<VisaCard121Status, string> = {
+  [VisaCard121Status.Active]: $localize`:@@debit-card-status-active:Active`,
+  [VisaCard121Status.Issued]: $localize`:@@debit-card-status-issued:Issued`,
+  [VisaCard121Status.Blocked]: $localize`:@@debit-card-status-blocked:Blocked`,
+  [VisaCard121Status.Paused]: $localize`:@@debit-card-status-paused:Paused`,
+  [VisaCard121Status.SuspectedFraud]: $localize`:@@debit-card-status-suspected-fraud:Suspected fraud`,
+  [VisaCard121Status.Unknown]: $localize`:@@debit-card-status-unknown:Unknown`,
+  [VisaCard121Status.Substituted]: $localize`:@@debit-card-status-substituted:Substituted`,
+  [VisaCard121Status.CardDataMissing]: $localize`:@@debit-card-status-card-data-missing:Debit card data missing`,
+};
