+++ interfaces/portal/src/app/domains/fsp-account-management/intersolve-visa.model.ts
@@ -1,0 +1,5 @@
+import { IntersolveVisaWalletDto } from '@121-service/src/fsp-integrations/integrations/intersolve-visa/dtos/internal/intersolve-visa-wallet.dto';
+
+import { Dto } from '~/utils/dto-type';
+
+export type WalletWithCards = Dto<IntersolveVisaWalletDto>;
