+++ interfaces/portal/src/app/domains/registration/registration.helper.ts
@@ -1,5 +1,4 @@
 import { ActivityTypeEnum } from '@121-service/src/activities/enum/activity-type.enum';
-import { VisaCard121Status } from '@121-service/src/fsp-integrations/integrations/intersolve-visa/enums/wallet-status-121.enum';
 import { DuplicateStatus } from '@121-service/src/registration/enum/duplicate-status.enum';
 import { RegistrationStatusEnum } from '@121-service/src/registration/enum/registration-status.enum';
 
@@ -43,17 +42,6 @@ export const REGISTRATION_STATUS_VERB: Record<RegistrationStatusEnum, string> =
 export const DUPLICATE_STATUS_LABELS: Record<DuplicateStatus, string> = {
   [DuplicateStatus.duplicate]: $localize`:@@duplicate-status-duplicate:Duplicate`,
   [DuplicateStatus.unique]: $localize`:@@duplicate-status-unique:Unique`,
-};
-
-export const VISA_CARD_STATUS_LABELS: Record<VisaCard121Status, string> = {
-  [VisaCard121Status.Active]: $localize`:@@debit-card-status-active:Active`,
-  [VisaCard121Status.Issued]: $localize`:@@debit-card-status-issued:Issued`,
-  [VisaCard121Status.Blocked]: $localize`:@@debit-card-status-blocked:Blocked`,
-  [VisaCard121Status.Paused]: $localize`:@@debit-card-status-paused:Paused`,
-  [VisaCard121Status.SuspectedFraud]: $localize`:@@debit-card-status-suspected-fraud:Suspected fraud`,
-  [VisaCard121Status.Unknown]: $localize`:@@debit-card-status-unknown:Unknown`,
-  [VisaCard121Status.Substituted]: $localize`:@@debit-card-status-substituted:Substituted`,
-  [VisaCard121Status.CardDataMissing]: $localize`:@@debit-card-status-card-data-missing:Debit card data missing`,
 };
 
 export const ACTIVITY_LOG_ITEM_TYPE_LABELS: Record<ActivityTypeEnum, string> = {
